+++ app/templates/components/page-nav.hbs
@@ -1,0 +1,4 @@
+<article class="page-nav">
+  <a href="#" class="left btn"  {{action "prevPage"}}>&lt; Prev</a>
+  <a href="#" class="right btn" {{action "nextPage"}}>Next &gt;</a>
+</article>
